some changes on YCQL entities, YCQL transaction 2,3,4

+++ ycql/src/main/java/com/cs5424t/ycql/Entities/Customer.java
@@ -2,7 +2,6 @@ package com.cs5424t.ycql.Entities;
 
 import com.cs5424t.ycql.Entities.PrimaryKeys.CustomerPK;
 import lombok.Data;
-import org.springframework.data.annotation.Id;
 import org.springframework.data.cassandra.core.mapping.Column;
 import org.springframework.data.cassandra.core.mapping.PrimaryKey;
 import org.springframework.data.cassandra.core.mapping.Table;
@@ -65,7 +64,7 @@ public class Customer implements Serializable {
     private BigDecimal balance;
 
     @Column("c_ytd_payment")
-    private Double ytdPayment;
+    private Float ytdPayment;
 
     @Column("c_payment_cnt")
     private Integer numOfPayment;
